VCV-150: Refactor auth files
