+++ src/api/auth/validation/signup-schema.ts
