+++ tsconfig.json
@@ -23,8 +23,6 @@
         "paths": {
             "@/*": ["./src/*"],
             "@/features/*": ["./src/features/*"],
-            "@/shared/*": ["./src/shared/*"],
-            "@/ui/*": ["./src/shared/ui/*"]
         },
         "verbatimModuleSyntax": true
     },
